# OAS/Ingredient.yaml
openapi: 3.1.0
info:
  title: Ingredient
  description: Ingredient
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
paths:
  /ingredients:
    get:
      tags:
        - Ingredient
      summary: Récupère tous les ingrédients
      description: Retourne la liste des ingredients
      responses:
        '200':
          description: Liste des ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
        '204':
          description: Aucun ingrédient trouvé
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: string
                example: "Erreur au niveau du serveur"

  /ingredients/{id}:
    get:
      tags:
        - Ingredient
      summary: Récupérer un ingrédient par son ID
      description: Retourne un ingrédient selon son ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID de l'ingrédient
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Ingrédient trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientDTO'
        '404':
          description: Ingrédient non trouvé
          content:
            application/json:
              schema:
                type: string
                example: "Ingredient avec cette n'existe pas"
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: string
                example: "Erreur au niveau du serveur"

  /ingredients/{id}/stock:
    get:
      tags:
        - Ingredient
      summary: Récupérer la valeur du stock d'un ingrédient
      description: Retourne la valeur du stock d'un ingrédient à un instant donné
      parameters:
        - name: id
          in: path
          required: true
          description: ID de l'ingrédient
          schema:
            type: integer
            example: 1
        - name: at
          in: query
          required: false
          description: Instant auquel évaluer le stock (format ISO 8601)
          schema:
            type: string
            format: date-time
            example: "2024-01-15T10:30:00Z"
        - name: unit
          in: query
          required: false
          description: Unité de mesure du stock
          schema:
            $ref: '#/components/schemas/UnitType'
      responses:
        '200':
          description: Valeur du stock récupérée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockValue'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                type: string
                example: "Paramètres de requête invalides"
        '404':
          description: Ingrédient non trouvé
          content:
            application/json:
              schema:
                type: string
                example: "Ingredient id not found"
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: string
                example: "Erreur au niveau du serveur"

  /dishes:
    get:
      tags:
        - Dishes
      summary: Récupère la liste des plats
      description: Retourne la liste des plats
      responses:
        '200':
          description: Liste des plats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DishDTO'
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: string
                example: "Erreur au niveau du serveur"



components:
  schemas:
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: tomate
        price:
          type: number
          format: double
          example: 1200.0
        category:
          type: string
          example: "VEGETABLE"

    DishDTO:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Romazava"
        sellingPrice:
          type: number
          format: double
          example: 15000.0
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientDTO'

    IngredientDTO:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "tomate"
        category:
          type: string
          example: "VEGETABLE"
        price:
          type: number
          format: double
          example: 1200.0

    StockValue:
      type: object
      properties:
        quantity:
          type: number
          format: double
          example: 10.0
        unit:
          $ref: '#/components/schemas/UnitType'

    UnitType:
      type: string
      enum:
        - PCS
        - KG
        - L
      example: KG